requirements.txt
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
apache-airflow==2.10.4
mlflow==2.18.0
matplotlib==3.9.2
seaborn==0.13.2
joblib==1.4.2
python-dotenv==1.0.1
pytest==8.3.3
